+++ package.json
@@ -8,7 +8,8 @@
 	},
 	"dependencies": {
 		"react": "16.0.0",
-		"react-native": "0.50.1"
+		"react-native": "0.50.1",
+		"react-navigation": "^1.0.0-beta.19"
 	},
 	"devDependencies": {
 		"babel-jest": "21.2.0",
